pyproject.toml
[project]
name = "Telegram-aggregator-bot"
version = "0.1.0"
description = "A Telegram userbot for aggregating and managing bulk chat messages."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiocron>=2.1",
    "beanie>=2.0.0",
    "dotenv>=0.9.9",
    "motor>=3.7.1",
    "pydantic-settings>=2.10.1",
]

[dependency-groups]
dev = [
    "ruff>=0.12.11",
]
userbot = [
    "pyrogram>=2.0.106",
    "tgcrypto>=1.2.5",
]
